requirements.txt
Flask==2.1.2
pymongo==4.5.0
scikit-learn==1.2.0
numpy==1.24.2
pandas==1.5.3
paho-mqtt==1.6.1
